requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
attrs==25.3.0
banks==2.2.0
beautifulsoup4==4.13.4
blinker==1.9.0
boto3==1.39.16
botocore==1.39.16
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
crewai==0.5.0
cycler==0.12.1
dataclasses-json==0.6.7
deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
eval-type-backport==0.2.2
exceptiongroup==1.3.0
fastapi==0.116.1
filetype==1.2.0
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2025.7.0
gitdb==4.0.12
gitpython==3.1.45
greenlet==3.2.3
griffe==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib-resources==6.5.2
jinja2==3.1.6
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.7
langchain==0.1.0
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.2.post1
langsmith==0.0.87
llama-cloud==0.1.35
llama-cloud-services==0.6.53
llama-index==0.12.32
llama-index-agent-openai==0.4.8
llama-index-cli==0.4.1
llama-index-core==0.12.32
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.8.0
llama-index-llms-openai==0.3.38
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.8
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.53
lxml==6.0.0
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.4
multidict==6.6.3
mypy-extensions==1.1.0
narwhals==2.0.1
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.9.1
numpy==1.26.4
openai==1.98.0
packaging==23.2
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
propcache==0.3.2
protobuf==6.31.1
pyarrow==17.0.0
pydantic==2.11.7
pydantic-core==2.33.2
pydeck==0.9.1
pymupdf==1.26.3
pyparsing==3.2.3
pypdf==5.9.0
pypdf2==3.0.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-pptx==1.0.2
pytz==2025.2
pyyaml==6.0.2
referencing==0.36.2
regex==2025.7.33
requests==2.32.4
rpds-py==0.26.0
s3transfer==0.13.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
sqlalchemy==2.0.42
starlette==0.47.2
streamlit==1.47.1
striprtf==0.0.26
tenacity==8.5.0
tiktoken==0.5.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.35.0
wrapt==1.17.2
xlsxwriter==3.2.5
yarl==1.20.1
zipp==3.23.0
